src/workstation_orchestrator/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o src/workstation_orchestrator/requirements.txt
annotated-types==0.7.0
    # via pydantic
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via workstation_orchestrater (pyproject.toml)
distro==1.9.0
    # via workstation_orchestrater (pyproject.toml)
idna==3.10
    # via requests
pydantic==2.10.5
    # via workstation_orchestrater (pyproject.toml)
pydantic-core==2.27.2
    # via pydantic
pyyaml==6.0.2
    # via workstation_orchestrater (pyproject.toml)
requests==2.32.3
    # via workstation_orchestrater (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
